Move utilities code to separate module
